refactor: handle navigation logic in dapps explorer screen
