+++ src/dappsExplorer/DAppsExplorerScreenLegacy.tsx
@@ -89,7 +89,7 @@ export function DAppsExplorerScreenLegacy() {
   }, [featuredDapp])
 
   return (
-    <SafeAreaView style={styles.safeAreaContainer} edges={['top']}>
+    <SafeAreaView testID="DAppsExplorerScreenLegacy" style={styles.safeAreaContainer} edges={['top']}>
       <DrawerTopBar
         middleElement={<Text style={headerStyles.headerTitle}>{t('dappsScreen.title')}</Text>}
         rightElement={
