+++ src/navigator/DrawerNavigator.tsx
@@ -41,12 +41,9 @@ import ContactCircleSelf from 'src/components/ContactCircleSelf'
 import PhoneNumberWithFlag from 'src/components/PhoneNumberWithFlag'
 import { RewardsScreenOrigin } from 'src/consumerIncentives/analyticsEventsTracker'
 import {
-  dappsFilterEnabledSelector,
-  dappsListApiUrlSelector,
-  dappsSearchEnabledSelector,
+  dappsListApiUrlSelector
 } from 'src/dapps/selectors'
-import DAppsExplorerScreen from 'src/dappsExplorer/DAppsExplorerScreen'
-import DAppsExplorerScreenLegacy from 'src/dappsExplorer/DAppsExplorerScreenLegacy'
+import DAppsExplorerScreen from 'src/dappsExplorer/DAppsExplorerScreenFilter'
 import { fetchExchangeRate } from 'src/exchange/actions'
 import ExchangeHomeScreen from 'src/exchange/ExchangeHomeScreen'
 import WalletHome from 'src/home/WalletHome'
@@ -202,9 +199,6 @@ export default function DrawerNavigator() {
   const { t } = useTranslation()
   const isCeloEducationComplete = useSelector(celoEducationCompletedSelector)
   const dappsListUrl = useSelector(dappsListApiUrlSelector)
-  const dappsFilterEnabled = useSelector(dappsFilterEnabledSelector)
-  const dappsSearchEnabled = useSelector(dappsSearchEnabledSelector)
-  const shouldShowDappsExplorerLegacy = !(dappsFilterEnabled || dappsSearchEnabled)
 
   const shouldShowRecoveryPhraseInSettings = useSelector(shouldShowRecoveryPhraseInSettingsSelector)
   const backupCompleted = useSelector(backupCompletedSelector)
@@ -272,9 +266,7 @@ export default function DrawerNavigator() {
       {!!dappsListUrl && (
         <Drawer.Screen
           name={Screens.DAppsExplorerScreen}
-          component={
-            shouldShowDappsExplorerLegacy ? DAppsExplorerScreenLegacy : DAppsExplorerScreen
-          }
+          component={DAppsExplorerScreen}
           options={{
             title: t('dappsScreen.title'),
             drawerIcon: DappsExplorer,
